Moved string2pos on server side
